+++ CHANGELOG.md
@@ -7,10 +7,15 @@
 - Added support for function extensions.
 - Added the built-in `length()` function.
 - Added the built-in `count()` function. `count()` is an alias for `length()`
-- Added the built-in `keys()` function (non-standard).
 - Support filters without parentheses.
 - Adhere to IETF JSONPath draft escaping in quoted property selectors.
 - Handle UTF-16 surrogate pairs in quoted property selectors.
+
+**Features**
+
+- Added the built-in `keys()` function.
+- Added `parent` and `children` properties to `JSONPathMatch`. Now we can traverse the "document tree" after finding matches.
+- Added a `parts` property to `JSONPathMatch`. `parts` is a tuple of `int`s, `slice`s and `str`s that can be used with `JSONPathEnvironment.getitem()` to get the matched object from the original data structure, or equivalent data structures. It is the keys, indices and slices that make up a concrete path.
 
 ## Version 0.2.0